Adding backtrace support
The desired effect is to have all fault exception calls print the call
stack so the user can determine what sequence of function calls caused
the fault to occur.

The problem is that the stack unwinding will only show the fault handler
and itself as the called functions, which is not very useful.

Removing unnecessary keep-folder.txt files